Added naive solution to reader-writer problem
Added a naive implementation of a solution to the reader-writer problem
This solution treats all readers as one "expensive" writer. The
implementation is simple, but there is a chance that real writer
threads will be starved.

Tested by adding a simple "checksum" of a shared buffer, and validating
the checksum by each reader thread.
Verified that the checksum is invalid when the synchronization logic
is removed, and that they remain valid in its current committed state.