Added timed mutex to show writer starvation
Changed mutex to timed mutex so an attempt to lock it can timeout after a specified period.
This will allow us to determine if the writer "starved" if the readers did not ever release
their lock on the resource. In these cases, the writer thread exits.

Added thread exit status checking at the end of the program.